+++ server/src/main/java/net/feedbacky/app/data/board/changelog/reaction/ChangelogReaction.java
@@ -1,0 +1,50 @@
+package net.feedbacky.app.data.board.changelog.reaction;
+
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import net.feedbacky.app.data.board.changelog.Changelog;
+import net.feedbacky.app.data.user.User;
+
+import org.hibernate.annotations.LazyToOne;
+import org.hibernate.annotations.LazyToOneOption;
+
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+
+import java.io.Serializable;
+
+/**
+ * @author Plajer
+ * <p>
+ * Created at 20.12.2021
+ */
+@Entity
+@Table(name = "boards_changelog_reactions")
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@EqualsAndHashCode(onlyExplicitlyIncluded = true)
+public class ChangelogReaction implements Serializable {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @EqualsAndHashCode.Include
+  private Long id;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @LazyToOne(LazyToOneOption.NO_PROXY)
+  private Changelog changelog;
+  @ManyToOne(fetch = FetchType.LAZY)
+  private User user;
+  private String reactionId;
+
+}