+++ server/src/main/java/net/feedbacky/app/data/board/dto/changelog/FetchChangelogDto.java
@@ -3,9 +3,12 @@ package net.feedbacky.app.data.board.dto.changelog;
 import lombok.Getter;
 import net.feedbacky.app.data.FetchResponseDto;
 import net.feedbacky.app.data.board.changelog.Changelog;
+import net.feedbacky.app.data.board.dto.changelog.reaction.FetchChangelogReactionDto;
 import net.feedbacky.app.data.user.dto.FetchSimpleUserDto;
 
 import java.util.Date;
+import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * @author Plajer
@@ -20,6 +23,7 @@ public class FetchChangelogDto implements FetchResponseDto<FetchChangelogDto, Ch
   private String description;
   private boolean edited;
   private FetchSimpleUserDto creator;
+  private List<FetchChangelogReactionDto> reactions;
   private Date creationDate;
 
   @Override
@@ -32,6 +36,7 @@ public class FetchChangelogDto implements FetchResponseDto<FetchChangelogDto, Ch
     this.description = entity.getDescription();
     this.edited = entity.isEdited();
     this.creator = new FetchSimpleUserDto().from(entity.getCreator());
+    this.reactions = entity.getReactions().stream().map(r -> new FetchChangelogReactionDto().from(r)).collect(Collectors.toList());
     this.creationDate = entity.getCreationDate();
     return this;
   }