+++ server/src/main/java/net/feedbacky/app/service/board/changelog/BoardChangelogService.java
@@ -3,7 +3,9 @@ package net.feedbacky.app.service.board.changelog;
 import net.feedbacky.app.data.board.dto.changelog.FetchChangelogDto;
 import net.feedbacky.app.data.board.dto.changelog.PatchChangelogDto;
 import net.feedbacky.app.data.board.dto.changelog.PostChangelogDto;
+import net.feedbacky.app.data.board.dto.changelog.reaction.FetchChangelogReactionDto;
 import net.feedbacky.app.data.idea.dto.FetchIdeaDto;
+import net.feedbacky.app.data.idea.dto.comment.reaction.FetchCommentReactionDto;
 import net.feedbacky.app.util.PaginableRequest;
 
 import org.springframework.http.ResponseEntity;
@@ -23,9 +25,13 @@ public interface BoardChangelogService {
 
   ResponseEntity<FetchChangelogDto> post(String discriminator, PostChangelogDto dto);
 
+  FetchChangelogReactionDto postReaction(long id, String reactionId);
+
   FetchChangelogDto patch(long id, PatchChangelogDto dto);
 
   ResponseEntity delete(long id);
+
+  ResponseEntity deleteReaction(long id, String reactionId);
 
   enum SortType {
     NEWEST, OLDEST